refactor cli tests
